bot-controlled: rename the authorization scheme
