+++ src/server/github.rs
@@ -1,6 +1,5 @@
 use errors::*;
 use reqwest::{header, Client, Method, RequestBuilder, StatusCode};
-use server::auth::Token;
 use server::tokens::Tokens;
 
 pub struct GitHubApi {
@@ -24,9 +23,7 @@ impl GitHubApi {
         };
 
         let mut req = self.client.request(method, &url);
-        req.header(header::Authorization(Token {
-            token: self.token.clone(),
-        }));
+        req.header(header::Authorization(format!("token {}", self.token)));
         req
     }
